feat - implement ExchangeRateUPdateHnadler
