+++ src/main/java/co/invest72/config/MoneyConfig.java
@@ -4,7 +4,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 
-import co.invest72.exchange_rate.domain.ExchangeRateRepository;
+import co.invest72.exchange_rate.application.ExchangeRateUpdateHandler;
 import co.invest72.exchange_rate.domain.service.Bank;
 import co.invest72.exchange_rate.domain.service.ExchangeRateService;
 import co.invest72.exchange_rate.infrastructure.api.FixedExchangeRateProvider;
@@ -19,7 +19,7 @@ public class MoneyConfig {
 	@Bean
 	@Profile(value = {"local", "test"})
 	public FixedExchangeRateProvider fixedExchangeRateProvider(ExchangeRateService exchangeRateService,
-		ExchangeRateRepository exchangeRateRepository) {
-		return new FixedExchangeRateProvider(exchangeRateService, exchangeRateRepository);
+		ExchangeRateUpdateHandler exchangeRateUpdateHandler) {
+		return new FixedExchangeRateProvider(exchangeRateService, exchangeRateUpdateHandler);
 	}
 }
